+++ NEWS.md
@@ -27,6 +27,9 @@
 
 ## Changes
 
+* The `plot()` method gets a `dot_shape` argument, to change the shape of data
+  points when `show_data = TRUE`.
+
 * `test_predictions` gains a `test_args` argument, to optionally pass further
   options to `test` for *emmeans* engine/options.